// src/types.ts
export interface components {
    schemas: {
// ...
        tagObject: {
            /**
             * Format: int32
             * @description Unique identifier for the tag
             */
            id: number;
            /** @description Name of the tag */
            name: string;
            /** @description Description of the tag */
            description: string | null;
            /** @description The text color of the tag */
            text_color: string | null;
            /** @description The background color of the tag */
            background_color: string | null;
            /**
             * Format: date-time
             * @description The date and time of when the tag was last updated (in the ISO 8601 extended format).
             */
            updated_at: string;
            /**
             * Format: date-time
             * @description The date and time of when the tag was created (in the ISO 8601 extended format).
             */
            created_at: string;
            /** @description If `true`, the tag will not show up when creating or updating transactions in the Lunch Money app. **Can it be assigned via the API** */
            archived: boolean;
            /**
             * Format: date-time
             * @description The date and time of when the tag was last archived or `null` if not archived
             */
            archived_at: string | null;
        };
// ...
    };
// ...
}